Refactor api.py: extract Providers, Models, and Bot Tokens routes
Continues the api.py split. Stacks on top of #363.

- Global providers (4 endpoints), model registry (5 endpoints), and
  global bot tokens (4 endpoints) → src/pinky_daemon/routes/providers.py

All 13 endpoints share `agents: AgentRegistry` as their only
dependency and were defined contiguously in the source. Same
APIRouter + set_dependencies(...) pattern as the previous
extractions.

`_provider_row_to_dict` moved into the new module (used only by /providers
routes). `_provider_public_row_to_dict` stayed in api.py because
/settings/default-provider also calls it.

api.py: 9,420 → 9,179 lines (-241, -2.6%)
Cumulative since main: 11,138 → 9,179 (-1,959 lines, -17.6%)

Tests: 1542 passed, 1 skipped (federation tests need nacl, pre-existing).
